overhaul

+++ database/migrations/2021_08_14_214333_create_token_price_table.php
@@ -15,7 +15,7 @@ class CreateTokenPriceTable extends Migration
     {
         Schema::create('token_price', function (Blueprint $table) {
             $table->id();
-            $table->double('price', 8, 2);
+            $table->float('price', 10, 4);
             $table->timestamps();
         });
     }
